Move the content of backends/bigquery/api.py to backends/bigquery/__init__.py Related to issue #2417
